Made some changes to the Promela code printer.
Changed the way timeouts are handled in ToSBit: they now apply only to the backend, and the frontend can try to recover however it sees fit.  In the case of ToPSbit, this means generating Promela code for the user to verify.

Also fixed a bug in the FEReplacer related to managing structs that was causing the latter semantic checker to fail.