Optimized the way for loops are partially evaluated. Makes a big difference in performance.
Introduced a new kind of assert (super assert) that is independent of the calling context. Most asserts are not super asserts, but I added a parameter to the constructor to make the type of assert explicit.

In order to make this possible, I made a small change to make function call boundaries explicit during partial evaluation.